+++ Sources/MudsnoteCore/NoteStore.swift
@@ -96,7 +96,7 @@ struct NoteSearchIndexEntry: Codable {
 }
 
 struct NoteSearchIndexDiskCache: Codable {
-    static let currentSchemaVersion = 2
+    static let currentSchemaVersion = 3
 
     let schemaVersion: Int
     let snapshot: NoteSearchIndexSnapshot